Rename defenitions to definitions
