+++ app/components/user/create/UserCreateForm.tsx
@@ -1,6 +1,6 @@
 "use client";
 import dynamic from "next/dynamic";
-import { Company } from "@/app/users/create/defenitions";
+import { Company } from "@/app/users/create/definitions";
 import { useActionState } from "react";
 import { createUser } from "@/app/users/create/actions";
